pyproject.toml
[tool.poetry]
name = "handbook"
version = "0.1.0"
description = "Programming Course for Engineers"
authors = ["Yves Chevallier <yves.chevallier@heig-vd.ch>"]
package-mode = false
license = "CC BY-NC-SA 4.0"
readme = "README.md"
keywords = ["mkdocs", "python", "academic", "course"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: MkDocs",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: French",
    "Programming Language :: C",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.4",
    "Programming Language :: Python :: 3.5",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python",
    "Topic :: Education",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development",
]

[tool.poetry.dependencies]

cairosvg = "^2.8.2"
chalk-diagrams = "^0.2.2"
deepmerge = "^2.0"
inflection = "^0.5.1"
markdown-exec = "^1.11.0"
mike = "^2.1.3"
mkdocs = "^1.6.1"
mkdocs-autorefs = "^1.4.3"
mkdocs-awesome-pages-plugin = "^2.10.1"
mkdocs-caption = "^1.3.0"
mkdocs-charts-plugin = "^0.0.13"
mkdocs-codeinclude-plugin = "^0.2.1"
mkdocs-drawio = { git = "https://github.com/yves-chevallier/mkdocs-drawio.git", branch = "testounet" }
mkdocs-drawio-exporter = "^0.10.2"
mkdocs-enumerate-headings-plugin = "^0.6.2"
mkdocs-epigraph = "^0.5.1"
mkdocs-extract-listings-plugin = "^0.2.1"
mkdocs-get-deps = "^0.2.0"
mkdocs-git-authors-plugin = "^0.10.0"
mkdocs-git-committers-plugin-2 = "^2.5.0"
mkdocs-git-revision-date-localized-plugin = "^1.4.7"
mkdocs-glightbox = "^0.5.1"
mkdocs-htmlproofer-plugin = "^1.3.0"
mkdocs-jupyter = "^0.25.1"
mkdocs-material = { git = "https://github.com/heig-tin-info/mkdocs-material-insiders.git" }
mkdocs-minify-plugin = ">=0.8.0"
mkdocs-pills = "^0.1.0"
mkdocs-plugin-exercises = { git = "https://github.com/yves-chevallier/mkdocs-exercise.git" }
mkdocs-snippets = "^1.3.2"
mkdocs-toggle-sidebar-plugin = "^0.0.8"
pillow = "^11.3.0"
#pillow-avif-plugin = ">=1.4.6"
pngquant = "^3.0.3"
pydantic = "^2.12.0"
pypdf = "^6.1.1"
python = "^3.10"
ruamel-yaml = "^0.18.15"
unidecode = "^1.4.0"
voluptuous = "^0.15.2"
wikipedia-api = "^0.8.1"

[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
ruff = "^0.14.0"
ipdb = "^0.13.13"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
preview = true

[tool.ruff.lint]
ignore = ["F401", "F403", "F405"]
select = ["E", "F", "I"]
fixable = ["I"]

[tool.pylint.main]
extension-pkg-allow-list = ["lxml", "lxml.etree"]
